requirements.txt
cgi-tools
pybase64
pycrypto
urllib3

